backend/requirements.txt
fastapi>=0.110.0,<0.125.0
uvicorn[standard]>=0.30.0,<0.35.0
openai>=1.17.0,<2.0.0
sqlmodel>=0.0.16,<0.1.0
sqlalchemy[asyncpg]>=2.0.0,<2.1.0
pgvector>=0.2.5,<0.3.0
asyncpg>=0.29.0,<0.30.0
redis>=5.0.0,<6.0.0
slowapi>=0.1.9,<0.2.0
pypdf>=4.0.0,<5.0.0
tiktoken>=0.6.0,<0.8.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.9,<0.1.0
python-dotenv>=1.0.0,<2.0.0
alembic>=1.13.0,<2.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<0.24.0
pytest-cov>=4.1.0,<5.0.0
httpx>=0.27.0,<0.28.0
ruff>=0.3.0,<0.8.0
websockets>=12.0,<13.0
pydantic-settings>=2.0.0,<3.0.0
